manifest creation ops, fixes and refactor
1.5.0

- added: button to create manifest.toml text from bl_info in opened text file
- fixed: bad regex on one search/replace string
- changed: code refactor
